+++ src/services/logger/logger.interface.ts
@@ -10,7 +10,7 @@ export interface Logger
      * @param messageTemplate - Message template for the log event.
      * @param properties - Properties that can be used to render the message template.
      */
-    fatal(messageTemplate: string, ...properties: any[]);
+    fatal(messageTemplate: string, ...properties: any[]): void;
 
     /**
      * Logs an event with the LogEventLevel.fatal severity.
@@ -18,14 +18,14 @@ export interface Logger
      * @param messageTemplate - Message template for the log event.
      * @param properties - Properties that can be used to render the message template.
      */
-    fatal(error: Error, messageTemplate: string, ...properties: any[]);
+    fatal(error: Error, messageTemplate: string, ...properties: any[]): void;
 
     /**
      * Logs an event with the LogEventLevel.error severity.
      * @param messageTemplate - Message template for the log event.
      * @param properties - Properties that can be used to render the message template.
      */
-    error(messageTemplate: string, ...properties: any[]);
+    error(messageTemplate: string, ...properties: any[]): void;
 
     /**
      * Logs an event with the LogEventLevel.error severity.
@@ -33,14 +33,14 @@ export interface Logger
      * @param messageTemplate - Message template for the log event.
      * @param properties - Properties that can be used to render the message template.
      */
-    error(error: Error, messageTemplate: string, ...properties: any[]);
+    error(error: Error, messageTemplate: string, ...properties: any[]): void;
 
     /**
      * Logs an event with the LogEventLevel.warning severity.
      * @param messageTemplate - Message template for the log event.
      * @param properties - Properties that can be used to render the message template.
      */
-    warn(messageTemplate: string, ...properties: any[]);
+    warn(messageTemplate: string, ...properties: any[]): void;
 
     /**
      * Logs an event with the LogEventLevel.warning severity.
@@ -48,14 +48,14 @@ export interface Logger
      * @param messageTemplate - Message template for the log event.
      * @param properties - Properties that can be used to render the message template.
      */
-    warn(error: Error, messageTemplate: string, ...properties: any[]);
+    warn(error: Error, messageTemplate: string, ...properties: any[]): void;
 
     /**
      * Logs an event with the LogEventLevel.information severity.
      * @param messageTemplate - Message template for the log event.
      * @param properties - Properties that can be used to render the message template.
      */
-    info(messageTemplate: string, ...properties: any[]);
+    info(messageTemplate: string, ...properties: any[]): void;
 
     /**
      * Logs an event with the LogEventLevel.information severity.
@@ -63,14 +63,14 @@ export interface Logger
      * @param messageTemplate - Message template for the log event.
      * @param properties - Properties that can be used to render the message template.
      */
-    info(error: Error, messageTemplate: string, ...properties: any[]);
+    info(error: Error, messageTemplate: string, ...properties: any[]): void;
 
     /**
      * Logs an event with the LogEventLevel.debug severity.
      * @param messageTemplate - Message template for the log event.
      * @param properties - Properties that can be used to render the message template.
      */
-    debug(messageTemplate: string, ...properties: any[]);
+    debug(messageTemplate: string, ...properties: any[]): void;
 
     /**
      * Logs an event with the LogEventLevel.debug severity.
@@ -78,14 +78,14 @@ export interface Logger
      * @param messageTemplate - Message template for the log event.
      * @param properties - Properties that can be used to render the message template.
      */
-    debug(error: Error, messageTemplate: string, ...properties: any[]);
+    debug(error: Error, messageTemplate: string, ...properties: any[]): void;
 
     /**
      * Logs an event with the LogEventLevel.verbose severity.
      * @param messageTemplate - Message template for the log event.
      * @param properties - Properties that can be used to render the message template.
      */
-    verbose(messageTemplate: string, ...properties: any[]);
+    verbose(messageTemplate: string, ...properties: any[]): void;
 
     /**
      * Logs an event with the LogEventLevel.verbose severity.
@@ -93,5 +93,5 @@ export interface Logger
      * @param messageTemplate - Message template for the log event.
      * @param properties - Properties that can be used to render the message template.
      */
-    verbose(error: Error, messageTemplate: string, ...properties: any[]);
+    verbose(error: Error, messageTemplate: string, ...properties: any[]): void;
 }